pyproject.toml
[project]
name = "mcp-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.86.0",
    "djangorestframework>=3.17.1",
    "drf-spectacular>=0.29.0",
    "httpx>=0.28.1",
    "mcp[cli]>=1.26.0",
    "python-dotenv>=1.2.2",
    "requests>=2.33.0",
]

[dependency-groups]
dev = [
    "ruff>=0.15.8",
]
